Move Deadline.java, Event.java, Parser.java, Storage.java, Task.java, TaskList.java, Todo.java, Ui.java to PacMan folder and bundle them to packages
