+++ src/main/java/PacMan/Event.java
@@ -1,3 +1,5 @@
+package PacMan;
+
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDate;
