+++ src/main/java/PacMan/Storage.java
@@ -1,3 +1,5 @@
+package PacMan;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.util.ArrayList;
