+++ src/main/java/PacMan/TaskList.java
@@ -1,3 +1,5 @@
+package PacMan;
+
 import java.util.ArrayList;
 
 public class TaskList {
